+++ CHANGELOG.md
@@ -7,13 +7,18 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## 0.4.0 - Unreleased
 
+## Fixed
+
+- Don't trust MSB of header/trailer without ECC
+- Fix panic when running RS on the header with 11 errors
+
 ## Added
 
 - "--test-simd" command to test available instruction set of the CPU
 
 ## Changed
 
-- Removed cargo/config.toml to achieve maximal compatibility with x86_64 CPUs.
+- Removed cargo/config.toml to achieve maximal compatibility with x86_64 CPUs
 
 ## 0.3.0 - 2025-08-31