+++ libtoa/CHANGELOG.md
@@ -5,6 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## 0.4.0 - Unreleased
+
+## Fixed
+
+- Don't trust MSB of header/trailer without ECC
+- Fix panic when running RS on the header with 11 errors
+
 ## 0.3.0 - 2025-08-31
 
 ## Changes